+++ src/TelegramBot.h
@@ -59,8 +59,8 @@ public:
 
     void getMe();
     void sendMessage(int64_t chatId, const std::string& text, ParseMode parseMode = ParseMode::None, bool disableWebPagePreview = false, bool disableNotification = false, int32_t replyToMessageId = 0, const TgBot::GenericReply::Ptr replyMarkup = TgBot::GenericReply::Ptr());
-    void sendLocation(int64_t chatId, float latitude, float longitude, bool disableNotification = false, int32_t replyToMessageId = 0, const TgBot::GenericReply::Ptr replyMarkup = TgBot::GenericReply::Ptr());
-    void sendVenue(int64_t chatId, float latitude, float longitude, const std::string& title, const std::string& address, const std::string& foursquareId = std::string(), bool disableNotification = false, int32_t replyToMessageId = 0, const TgBot::GenericReply::Ptr replyMarkup = TgBot::GenericReply::Ptr());
+    void sendLocation(int64_t chatId, double latitude, double longitude, bool disableNotification = false, int32_t replyToMessageId = 0, const TgBot::GenericReply::Ptr replyMarkup = TgBot::GenericReply::Ptr());
+    void sendVenue(int64_t chatId, double latitude, double longitude, const std::string& title, const std::string& address, const std::string& foursquareId = std::string(), bool disableNotification = false, int32_t replyToMessageId = 0, const TgBot::GenericReply::Ptr replyMarkup = TgBot::GenericReply::Ptr());
     void sendContact(int64_t chatId, const std::string& phoneNumber, const std::string& firstName, const std::string& lastName = std::string(), bool disableNotification = false, int32_t replyToMessageId = 0, const TgBot::GenericReply::Ptr replyMarkup = TgBot::GenericReply::Ptr());
